Update loader.c
I revised the additional GetMUI function to utilize the "GetUserDefaultLocaleName" function to acquire the proper locale name and subsequent location of the MUI file.  This allowed me to remove the slower homemade "switch/case" code I had added to attempt to acquire the locale name.  Earlier attempts to utilize the "GetUserDefaultLocaleName" function had not worked because that function also called the "FindResourceExW" function which in turn called the "GetMUI" function, setting up a recursive call spiral that ultimately resulted in a stack overflow.  Nesting the call to the "GetUserDefaultLocaleName" function inside a recursive condition test now avoids this issue and allows for a more compact and more efficient function.