Update loader.c
Revised the code enhancements utilizing the Wine development coding conventions for possible consideration to include this enhancement in a future iteration of the Wine application.  Also, this code enhancement has been applied to Wine version 5.15 and system tested.  Other than a bit of a time lag during the initial startup of an application with a "mui" file, the applications that utilize this enhancement appear to function properly.