Update loader.c
Per feedback from my Wine application contace, the proposed enhancement would not be accepted as is with a hard-coded range of values test.  Therefore, the "anecdotal range of values condition" was removed from the "LoadResource" function and the function test within the "FindResourceExW" function was simplified.  The only negative knock-on effect with this simplification is that potentially no language neutral resources stored within the executable file such as icons will be found and loaded.  But, this most likely will not result in making the application non-functional and is a small trade-off compared to the benefit of finding and loading resources such as menus and strings from the associated MUI file.