Update loader.c
Per correspondence with Wine application experts, I revised the loader program file further to make a second attempt to find a resource in an associated MUI file if the initial resource load attempt did not work.  Also, I was able to condition the acquisition of the MUI module handle in the resource loader function without utilizing constant value hard-coding and thus avoiding potential page faults.  With that, both the language neutral resources and the language specific resources all appear to be found and loaded.