Update loader.c
Reapplied the multi-language enhancements to the source code for version 5.17 of the loader.c file, and made a small change to the calling of the library load function within the "get_mui" function.  Instead of calling function "LoadLibraryW" which in turn calls function "LoadLibraryExW" with default parameters, a direct call to function "LoadLibraryExW" is made with the same default parameters.  This reduces the list of function calls and slightly reduces the addition and removal of functions from the call stack during the loading of resources, again making the enhancement a bit simpler and a bit more efficient.  Also, for better clarification, I revised the name of the other additional helper function to be "get_rsc_handle".