Update loader.c
The additional functions were moved next to the functions that were being enhanced so that the code changes and additions resided together in the same vicinity within the program file to clarify what enhancements had been made and make it easier to reapply these enhancements in future versions of the source code.  Also, this rearrangement of code enhancements makes it more conducive for possible submission as a program patch.