Create loader.c
I am saving a copy of the loader.c file for the stable version.  This copy will get updated when the next stable version of Wine comes out (probably next year).  This gives anyone interested in utilizing the multi-language resource retrieval enhancement for the stable version to find it here.